chargebee-processor/requirements.txt
./oep-core

aws-psycopg2
requests
